Category for tests using splittable DoFn
